update, split conrod and sprites
