+++ graphics/Cargo.toml
@@ -5,14 +5,14 @@ version = "0.1.0"
 authors = ["Thomas Koehler <basta.t.k+git@gmail.com>"]
 
 [dependencies]
-gfx = { git = "https://github.com/Bastacyclop/gfx", branch = "persistent" }
-gfx_core = { git = "https://github.com/Bastacyclop/gfx", path="src/core", branch = "persistent" }
-gfx_device_gl = { git = "https://github.com/Bastacyclop/gfx", path="src/backend/gl", branch = "persistent" }
-gfx_window_glutin = { git = "https://github.com/Bastacyclop/gfx", path="src/window/glutin", branch = "persistent" }
+gfx = { git = "https://github.com/gfx-rs/gfx" }
+gfx_core = { git = "https://github.com/gfx-rs/gfx" }
+gfx_device_gl = { git = "https://github.com/gfx-rs/gfx" }
+gfx_window_glutin = { git = "https://github.com/gfx-rs/gfx" }
 glutin = "0.6"
 rusttype = "0.2"
-conrod = { version = "0.39", features = ["glutin"] }
-cgmath = "0.10"
+conrod = { version = "0.41", features = ["glutin"] }
+cgmath = "0.11"
 image = "0.10"
 quickersort = "2.0"
 rayon = "0.4"
