+++ graphics/examples/conrod.rs
@@ -5,23 +5,21 @@ extern crate glutin;
 extern crate cgmath;
 extern crate rayon;
 
-use graphics::{Graphics, Camera};
-use graphics::combined::sprites::Renderer;
-use graphics::lights::*;
+use graphics::Graphics;
+use graphics::combined::conrod::Renderer;
 use graphics::types::ColorFormat;
+use conrod::widget;
 use glutin::{WindowBuilder, Event};
-use cgmath::{Point2, Vector2};
 use frameclock::*;
 
 fn main() {
     let builder = WindowBuilder::new()
         .with_title("Conrod".to_string())
         .with_dimensions(512, 512)
-        .with_gl(glutin::GlRequest::Specific(glutin::Api::OpenGl, (3, 2)))
-        .with_vsync();
+        .with_gl(glutin::GlRequest::Specific(glutin::Api::OpenGl, (3, 2)));
 
     let (window, mut graphics) = Graphics::new(builder);
-    let mut renderer = Renderer::new(&window, &mut graphics);
+    let mut renderer = Renderer::new(window.hidpi_factor(), &mut graphics);
 
     let mut ui = conrod::UiBuilder::new().build();
     ui.fonts.insert_from_file("resources/fonts/NotoSans/NotoSans-Regular.ttf").unwrap();
@@ -43,13 +41,7 @@ fn main() {
         (IMAGE_GREEN, texture.clone()),
         (IMAGE_BLUE, texture),
     };
-    
-    let camera = Camera::new(Point2::new(0., 0.), Vector2::new(1.0, 1.0), &window);
-    let ambient_light = AmbientLight {
-        color: LightColor::from_srgb([0.2, 0.2, 0.2]),
-        intensity: 0.5,
-    };
-    
+
     let mut frameclock = FrameClock::start(1.);
     let mut fps_counter = FpsCounter::new(1.);
     let mut mspf = 0.0;
@@ -78,10 +70,11 @@ fn main() {
         }
 
         ui.handle_event(conrod::event::render(delta_time, w, h, dpi_factor));
-        ui.set_widgets(|ref mut ui| {
+        {
+            let mut ui = &mut ui.set_widgets();
             use conrod::{Colorable, Positionable, Sizeable, Widget};
-            
-            conrod::Canvas::new().color(conrod::color::DARK_CHARCOAL).set(CANVAS, ui);
+
+            widget::Canvas::new().color(conrod::color::DARK_CHARCOAL).set(CANVAS, ui);
 
             let demo_text = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. \
                 Mauris aliquet porttitor tellus vel euismod. Integer lobortis volutpat bibendum. \
@@ -92,7 +85,7 @@ fn main() {
                 Quisque commodo nibh hendrerit nunc sollicitudin sodales. Cras vitae tempus ipsum. \
                 Nam magna est, efficitur suscipit dolor eu, consectetur consectetur urna.";
 
-            conrod::Text::new(demo_text)
+            widget::Text::new(demo_text)
                 .middle_of(CANVAS)
                 .w_of(CANVAS)
                 .font_size(20)
@@ -102,7 +95,7 @@ fn main() {
 
             let fps_text = &format!("{:.2} ms/frame - {} fps", mspf, fps);
 
-            conrod::Text::new(fps_text)
+            widget::Text::new(fps_text)
                 .top_left_with_margin_on(CANVAS, 6.)
                 .wh_of(CANVAS)
                 .font_size(14)
@@ -110,16 +103,16 @@ fn main() {
                 .align_text_left()
                 .set(FPS_TEXT, ui);
 
-            let style = conrod::LineStyle {
+            let style = widget::line::Style {
                 maybe_pattern: None,
                 maybe_color: None,
                 maybe_thickness: Some(20.),
-                maybe_cap: Some(conrod::LineCap::Round),
+                maybe_cap: Some(widget::line::Cap::Round),
             };
 
-            conrod::Line::abs_styled([0., 0.], [120., 0.], style).set(LINE, ui);
+            widget::Line::abs_styled([0., 0.], [120., 0.], style).set(LINE, ui);
 
-            let image = conrod::Image::new().w_h(64., 64.);
+            let image = widget::Image::new().w_h(64., 64.);
             image.clone()
                 .color(Some(conrod::color::RED))
                 .bottom_left_with_margin_on(CANVAS, 12.)
@@ -132,12 +125,11 @@ fn main() {
                 .color(Some(conrod::color::BLUE))
                 .right_from(IMAGE_GREEN, 12.)
                 .set(IMAGE_BLUE, ui);
-        });
-        
+        }
+
         let mut frame = graphics.draw();
-        //let conrod_data = ui.draw_if_changed().map(|p| (p, &image_map));
-        let conrod_data = Some((ui.draw(), &image_map));
-        renderer.submit_with_conrod(conrod_data, &camera, &ambient_light, &mut frame);
+        ui.draw_if_changed().map(|ps| renderer.changed(ps, &image_map, &mut frame));
+        renderer.render(&mut frame);
         frame.present(&window);
     }
 }
