+++ graphics/src/combined/mod.rs
@@ -1,3 +1,4 @@
 //! Combined renderers
 
+pub mod conrod;
 pub mod sprites;
