+++ graphics/src/specialized/mod.rs
@@ -5,4 +5,3 @@ pub mod glyphs;
 pub mod images;
 pub mod dynamic_lights;
 pub mod sprites;
-pub mod conrod;
